more search, cocktails not being indexed
